Index I think is finally done including mobile view to display everything correctly.
Worked on the nav bar more so it is now displaying more or less the way I want it to.

Blog has been revamped but still needs extra work to ensure that it looks good, still a bit off centre and the overall posts need styling and formating corretly still.

About page has some basic content, it will be quite a simple view overall so it doesn't need too much extra work on top of what I have done